refactor(ostd/la): move CPU context handling into separate module
